+++ engine/src/parser/expression_utils.cpp
@@ -90,6 +90,7 @@ bool is_binary_operator(operator_type op) {
 	case operator_type::BLZ_STR_REPLACE:
 	case operator_type::BLZ_TO_DATE:
 	case operator_type::BLZ_TO_TIMESTAMP:
+	case operator_type::BLZ_STR_TRIM:
 		assert(false);
 		// Ternary operator. Should not reach here
 		// Should be evaluated in place (inside function_evaluator_transformer) and removed from the tree
@@ -225,6 +226,7 @@ cudf::type_id get_output_type(operator_type op, cudf::type_id input_left_type, c
 	case operator_type::BLZ_STR_SUBSTRING:
 	case operator_type::BLZ_STR_REPLACE:
 	case operator_type::BLZ_STR_CONCAT:
+	case operator_type::BLZ_STR_TRIM:
 		return cudf::type_id::STRING;
 	case operator_type::BLZ_TO_DATE:
 		return cudf::type_id::TIMESTAMP_DAYS;
@@ -301,7 +303,8 @@ operator_type map_to_operator_type(const std::string & operator_token) {
 		{"REPLACE", operator_type::BLZ_STR_REPLACE},
 		{"TO_DATE", operator_type::BLZ_TO_DATE},
 		{"TO_TIMESTAMP", operator_type::BLZ_TO_TIMESTAMP},
-		{"||", operator_type::BLZ_STR_CONCAT}
+		{"||", operator_type::BLZ_STR_CONCAT},
+		{"TRIM", operator_type::BLZ_STR_TRIM},
 	};
 
 	if(OPERATOR_MAP.find(operator_token) == OPERATOR_MAP.end()){
@@ -592,6 +595,10 @@ std::string replace_calcite_regex(const std::string & expression) {
 	StringUtil::findAndReplaceAll(ret, "EXTRACT(FLAG(HOUR), ", "BL_HOUR(");
 	StringUtil::findAndReplaceAll(ret, "EXTRACT(FLAG(MINUTE), ", "BL_MINUTE(");
 	StringUtil::findAndReplaceAll(ret, "EXTRACT(FLAG(SECOND), ", "BL_SECOND(");
+	// Flag options for TRIM
+	StringUtil::findAndReplaceAll(ret, "TRIM(FLAG(BOTH),", "TRIM(\"BOTH\",");
+	StringUtil::findAndReplaceAll(ret, "TRIM(FLAG(LEADING),", "TRIM(\"LEADING\",");
+	StringUtil::findAndReplaceAll(ret, "TRIM(FLAG(TRAILING),", "TRIM(\"TRAILING\",");
 
 	StringUtil::findAndReplaceAll(ret, "/INT(", "/(");
 	return ret;